Remove all oc calls and allow for running from inside a cluster
Loosen the KUBECONFIG + ~/.kube/config check and only fail it if both
are missing *and* we're not running inside a pod in a cluster (i.e.
when KUBERNETES_SERVICE_HOST env var is set)

Also remove all remaining "oc" calls. Those are not really very ansible
*and* the json_path query we do in ansible is only available when using
the general.community collection which is not present in the ee
container we use by default.