Explicitely ignore ssh pubkeyacceptedalgorithms config option
When a user has an .ssh/config file that leverages recent options, like
pubkeyacceptedalgorithms, 'make install' using the wrapper will fail
with:

    Please make sure you have the correct access rights
    and the repository exists.
    /root/.ssh/config: line 643: Bad configuration option: pubkeyacceptedalgorithms

This is because my host has openssh-8.8p1-1.fc36.1.x86_64 whereas the container has
the older openssh-8.0p1-16.el8.x86_64 version which is not aware of the
above option.

Since pubkeyacceptedalgorithms is a common setting in newer ssh's that
allows you to connect to older and deprecated systems, let's explicitely
disable it.