Add python module implementing the push-secret functionality
This module rewrites the push_secrets.yaml functionality in python.
The reason for this rewrite is that it just became too complext to
manage in pure ansible and we really reached the limits of ansible and
it became an unreadable mess.

Also add an ansible-unittest makefile target to unittest the module.
For the time being we leave it out of CI until it is a bit more
refined/generic.