Split secret counter in two variables
The 'counter' variable is used to detect if it is the first vault kv
command to be run (i.e. vault kv put as opposed to vault kv patch for
subsequent runs), so we need to reset it for every top-level secret.
We introduce an additional variable to count every secret entry
uploaded.