Add "vault_path" to the parse_acm_secrets filter
This vault_path field will define the path to be used in vault for a
specific cluster: "hub" when the hub and the FQDN for all other nodes.

Doing it this way allows us to use "hub" as a fixed path to login to the
hub. I.e. we do not need to customize values-secret.yaml whenever we're
only using secrets in the hub and not per-cluster secrets.