Do not validate certs when we use letsencrypt on the endpoints
Due to https://issues.redhat.com/browse/ACM-4398 we cannot leave
validate_certs always on. Since we do not want to always disable it
as that would be a security regression, we simply disable it only
when the helm variable "letsencrypt.api_endpoint" is set to true.
In that case we know that we are going to use letsencrypt certificates
so it is okay to disable the verification until then.
Once/if the ACM bug above gets fixed we can drop this.

Tested on MCG